Tests should also use 'phantomjs' package for consistency.
The code has been updated to use the 'phantomjs' package to install and
locate the 'phantomjs' binary, but the tests has not.

If 'phantomjs' had install 'phantomjs' locally and 'phantomjs' was not
in $PATH, then the tests would fail, even though 'phantomjs' was
installed would be located by the code.

Now we test for phantomjs the same way the code would.