Generate tmp file names with a consistent method.
One one place, we were generating tmp files that included the PID, while
in another we were not.

This patch updates the code to use a shared _getTmpFile function to
enforce consistency here and clean-up the calling the function a bit.

Also, the 'tmp' variable was now the lone alias from the 'opts' object.
'opts.tmp' is refenced now instead to remove a bit of indirection.